update annealer cache management, parameters
the cache measurer was substantially inaccurate (i frankly cannot
understand why i wrote it the way i did. maybe it was right the first
time and there was some insight i forgot, but that seems unlikely). this
improves the measurement.

since the annealer's behavior was dependent on the cache measurements,
we need to adjust it's behavior. the new behavior is cache-independent,
with a lower initial temperature and a slower cooling schedule (from linear to 1-sqrt); this is
about the best balance i could find for reliability