update live sim in various ways
I did a bunch of work a few months back which I didn't commit because
the web UI build process is currently super annoying. Now there's a
bunch of interdependent changes and I am just committing them all at
once so that the repo doesn't fall hopelessly out of date. The web build
is still annoying but the actual web-facing stuff isn't in the repo yet
so uhh