fix inconsistencies wrt the order of strats
I ordered the strategies as (G, B, R) in some places and (R, G, B)
others. This notably resulted in the schedule moves focusing on a
strategy the live simulator never entered. Fixing this improves
search results, but not by as much as I expected; might simplify
schedule moves based on this information